Add last dequeuing by default

+++ lib/active_job/queue_adapters/delayed_job_adapter.rb
@@ -19,13 +19,17 @@ module ActiveJob
 
         def dequeue(job, *args, &blk)
           handler = job.new
-          Delayed::Job.all.each do |job_model|
+          all_jobs = Delayed::Job.all
+          return all_jobs.last.destroy if args.empty?
+          all_jobs.each do |job_model|
             enqueued_args = job_model.payload_object.args
             if handler.filter_job(enqueued_args, args, &blk)
               job_model.destroy
             end
           end
         end
+
+        def dequeue_empty()
       end
 
       class JobWrapper < Struct.new(:job, :args)
